Refactor code structure for improved readability and maintainability

+++ src/core/application/cases/rent/DeleteRentUseCase.ts
@@ -1,4 +1,5 @@
 import { IRentalRepository } from "@/core/domain/repositories/IRentalRepository";
+import { ServerError } from "@/utils/models/ServerError";
 
 export class DeleteRentUseCase {
   constructor(
@@ -6,6 +7,12 @@ export class DeleteRentUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
-    this.rentalRepo.delete(id);
+    const existingRent = await this.rentalRepo.find(id);
+    
+    if (!existingRent) {
+      throw new ServerError("Aluguel não encontrado ou já excluído.", 404);
+    }
+
+    await this.rentalRepo.delete(id);
   }
 }
